SP表示時のサイト内・ページ内ナビゲーションを追加
